Fix infinite scroll page height bug
When the initial page height was smaller than the window height, infinite
scroll did not work because there was no scroll bar, and thus a user
couldn't scroll down to trigger a scroll event that would load more content.

Now on page load or new pillar selection, a check is performed to see if
more content needs to be loaded initially, in case page height is smaller
than window height.

See: https://trello.com/c/npXA85Xs/739-5-414-l-as-a-user-i-should-not-have-to-click-load-more-for-more-content-to-appear-infinite-scroll